Moved all server dependent tests out of unit tests. Consolidated unit test targets
