--- Sources/ApolloSQLiteTestSupport/ApolloSQLiteTestSupport.h
@@ -1,11 +1,0 @@
-#import <Foundation/Foundation.h>
-
-//! Project version number for ApolloSQLiteTestSupport.
-FOUNDATION_EXPORT double ApolloSQLiteTestSupportVersionNumber;
-
-//! Project version string for ApolloSQLiteTestSupport.
-FOUNDATION_EXPORT const unsigned char ApolloSQLiteTestSupportVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <ApolloSQLiteTestSupport/PublicHeader.h>
-
-
