+++ Tests/Apollo-UnitTestPlan.xctestplan
@@ -1,0 +1,24 @@
+{
+  "configurations" : [
+    {
+      "id" : "36C7A583-06D6-4733-AF67-CC38D6FB14FF",
+      "name" : "Configuration 1",
+      "options" : {
+
+      }
+    }
+  ],
+  "defaultOptions" : {
+
+  },
+  "testTargets" : [
+    {
+      "target" : {
+        "containerPath" : "container:Apollo.xcodeproj",
+        "identifier" : "9FC7504D1D2A532D00458D91",
+        "name" : "ApolloTests"
+      }
+    }
+  ],
+  "version" : 1
+}
