+++ Tests/ApolloWebSocket-UnitTestPlan.xctestplan
@@ -1,0 +1,33 @@
+{
+  "configurations" : [
+    {
+      "id" : "02D5FC33-493B-45F5-9563-2FE5B7495F16",
+      "name" : "Configuration 1",
+      "options" : {
+
+      }
+    }
+  ],
+  "defaultOptions" : {
+    "codeCoverage" : {
+      "targets" : [
+        {
+          "containerPath" : "container:Apollo.xcodeproj",
+          "identifier" : "9B7BDA7C23FDE90400ACD198",
+          "name" : "ApolloWebSocket"
+        }
+      ]
+    },
+    "environmentVariableEntries" : [
+      {
+        "key" : "APOLLO_TEST_CACHE_PROVIDER",
+        "value" : "ApolloTestSupport.InMemoryTestCacheProvider"
+      }
+    ],
+    "threadSanitizerEnabled" : true
+  },
+  "testTargets" : [
+
+  ],
+  "version" : 1
+}
